Make contract non-ABI compliant
Since the ABI (and Beaker) does not yet have a way of handling variable length reference arrays, the distribute_donations method was broken up into 4 separate ones that each take 1, 2, 3, or 4 accounts and the same number of papers_scraped. This is very cumbersome and will make usage of the contract complicated. This commit changes it to a singular distribute_donations()void method that expects between 1-4 accounts and the same number of papers_scraped.